+++ recalc-vscode/src/frontend/plugin.ts
@@ -6,6 +6,7 @@ import { MenuController } from './controllers/menu.controller';
 import { MessageController } from './controllers/rpc.controller';
 
 import { DescriptionService } from './services/function-description.service';
+import { FunctionType } from '@univerjs/engine-formula';
 
 export class RecalcPlugin extends Plugin {
   static override type = UniverInstanceType.UNIVER_SHEET;
@@ -37,6 +38,20 @@ export class RecalcPlugin extends Plugin {
     this._injector.get(MenuController);
     this._injector.get(HoverController);
     this._injector.get(MessageController);
-    this._injector.get(IDescriptionService);
+    this._injector.get(IDescriptionService).registerDescriptions([
+      DescriptionService.functionDescription2Info({
+        name: "not",
+        type: FunctionType.Logical,
+        description: "Returns TRUE if the argument is FALSE, and FALSE if the argument is TRUE.",
+        abstract: "Logical negation.",
+        params: [
+          {
+            name: "logical",
+            detail: "A boolean value or expression to negate.",
+            example: "FALSE",
+          }
+        ]
+      })
+    ]);
   }
 }