+++ recalc-vscode/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Revision history for recalc-vscode
 
-## 0.0.23 -- 2025-03-06
+## 0.0.24 -- 2025-03-07
 
 * Initial version: create project structure for a custom editor.
 * Add generic rpc implementation, client implementation and logging facilities.
@@ -37,6 +37,8 @@
   spilling.
 * Add more end-to-end tests covering cell ranges.
 * Match new version of recalc + more tests.
+* Implement sheet-defined functions and description service (some fields are
+  still missing).
 
 <!-- References -->